Add restart confirmation

+++ setup.sh
@@ -11,4 +11,8 @@ for app in "${apps[@]}"; do
   source ./config/$app
 done
 
-sudo shutdown -r 0
+read -r -p "Do you want to restart? [y/N] " response
+
+if [[ "$response" =~ ^([yY][sS]|[yY])+$ ]]; then
+  sudo shutdown -r 0
+fi
